Uploaded
Scientists measure an object’s mass in kilograms and its weight in newtons. If you know an object’s mass, you can calculate its
weight in newtons with the following formula:
weight = mass * 9.8
Write a program that asks the user to enter an object’s mass, and then calculates and displays its weight. If the object weighs
more than 1000 newtons, display a message indicating that it is too heavy. If the object weighs less than 10 newtons, display a
message indicating that the object is too light.